Throw with a classname when the template is falsey
Previously if there was no template definition on a View, rendering the view would throw an exception where the last non-Backbone/Marionette function would be at level 17 in the backtrace. This made tracing the view without template a little annoying. Since a boolean check is not that much of a heavyweight (and the render call only occurs twice) we raise before calling Renderer.render() if we find a template that is falsey (undefined, false or null)