Update devDependencies
I've updated the devDependencies based on the output from
running the `npm outdated` command.

The only dependency I've left alone is on sinon. Going to
any version of sinon >1.10.3 breaks `behaviors.spec.js`
and I was not able to figure out exactly why as my experience
with sinon is very limited.

Wether you choose to merge or ignore this PR, at least now
you know ahead of time that there's an issue with it.